requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
python-decouple==3.8
